Updates to vehicle view, linked directly to create oil change, started link to oil change listing view per vehicle, updated database for vehicle id to be int instead of byte

+++ OilChangeTracker/Models/OilChangeEvent.cs
@@ -11,7 +11,7 @@ namespace OilChangeTracker.Models
         public Vehicle Vehicle{ get; set; }
 
         [Required]
-        public byte VehicleId { get; set; }
+        public int VehicleId { get; set; }
 
         [Required]
         public DateTime WorkDate { get; set; }
